Add SQL::NumericMethods#coerce, allowing code such as Sequel.expr{1 - x}
This fixes a long standing annoyance when using Sequel, especially
when using noncommunitive operators such as -, /, and now **.

Simply by defining this on NumericMethods, it also works on Symbols
when using the core extensions.  However, it does not work when
using the core refinements, since coerce is called implicitly and
not explicitly.