Support IntegerMigrator :relative option, for running only the specified number of migrations up or down
This makes it simple to migrate the last applied migration down
using :relative=>-1.

This only works for the IntegerMigrator, which has a fixed order,
it doesn't make sense for the TimestampMigrator, which allows for
out of order migrations.

Also, this handles migration targets that are below 0 or above
the latest migration version by capping the target to one of
those two extremes.