Add Sequel::Database.set_shared_adapter_scheme to allow external adapters to support the mock adapter
Previously, external adapters could not easily integrate with the
mock adapter.  This allows external adapters to just add some code
to their shared adapter that will cause the mock adapter to
recognize them.  The one requirement is that the mock adapter
module needs to call Sequel::Database.set_shared_adapter_scheme
to register the shared adapter.  Optionally, it can provide
a mock_adapter_setup singleton method, which is called with
the mock Database object.  Additionally, if shared adapter
contains a DatabaseMethods module, the mock Database is
extended with it, and if it contains a DatasetMethods module,
the mock Database's datasets are extended with it. Example:

class Sequel::MyDB
  Sequel::Database.set_shared_adapter_scheme :mydb, :self

  def self.mock_adapter_setup(db)
    # ...
  end

  module DatabaseMethods
    # ...
  end

  module DatasetMethods
    # ...
  end
end