Add Database#rollback_checker, which returns a proc that returns whether the in progress transaction is rolled back
This is designed to be used when many different objects need to
check whether the given transaction has been rolled back or
committed.  Previously, you could do this using an after_rollback
hook, but that required a separate hook for every object you
were interested in, and it kept the objects in memory unless
the transaction was completed.  With this API, objects can get
transaction status if they need it, but the transaction doesn't
have a reference to them, and therefore the objects can be
garbage collected before the transaction finishes.