Add block support to Sequel::Model::Errors#add
This is something I've been monkey patching in my projects, and figured it's possibly worth adding to Sequel.

Here's the logic. Often when validating fields, a single field may have many validations. For example you may first check the format is correct, and then check that the value is within range; there's no use checking if the value is within range if it's not even in the right format to begin with. The block provides a convenient mechanism for this common situation. Refer to the following contrived example:

``` ruby
errors.add(:start_date, "Start date must be a valid date") { DateTime === start_date }
errors.add(:start_date, "Start date must not be in the past") { start_date >= DateTime.today }
```

In the example above, the second validation is only checked is there's not already an error associated with the `start_date` field. The block defines the valid condition, so it it returns false, the validation has failed, and the error message is appended.