Add SQL::Identifier#[] and SQL::QualifiedIdentifier#[] for creating qualified identifiers
This offers a simpler syntax for creating qualified identifiers from
existing identifiers.  Currently, you have to use qualify, but that
looks kind of backwards as it puts the qualifer after the identifier
instead before.  This works well with Sequel#[], allowing syntax such
as Sequel[:table][:column] for table.column.

Change the documentation to start recommending Sequel[][] instead of
Sequel.qualify.  Most of the internal code that calls Sequel.qualify
remains the same, since Sequel.qualify is different in that it can
handle strings as the qualifier, while Sequel[] if given a string
will return a StringExpression and not an identifier.  In the cases
where Sequel.qualify used an explicit symbol, I switched it to
Sequel[][].