Add Sequel.split_symbols=, to support the disabling of splitting symbols with double/triple underscores
The splitting of symbols such as :table__column, :column___alias,
and :table__column___alias has been around since Sequel's earliest
commits, before it was even called Sequel.  At the time, it was
the only way to represent qualified or aliased identifiers.

Over the years, Sequel has grown additional ways to create
qualified and aliased identifiers, and it has gotten to the point
where the alternative ways are fairly easy to use, enough that
most developers could probably stand to use them instead of relying
on symbol splitting, and I plan to introduce extensions to make
it even easier to create qualified and/or aliased identifiers in the
future.

Symbol splitting was always a hack, and it has caused many, many
issues over the years.  There are databases in use that use
double and triple underscores in their identifiers, and it's
basically whack-a-mole to fix the bugs in such cases.  By
disabling symbol splitting, things are much more likely to
work correctly on such databases.

At some point in the future, I would like to disable symbol
splitting by default, and at some point after that remove it
entirely and clean up the codebase so it doesn't even attempt
symbol splitting.  I'm considering disabling symbol splitting
in Sequel 5, but I have not made a decision on that yet.

In terms of changes in this commit, the doc/lib changes were
made by hand, but the spec changes were made via a ruby program
using gsub and Sequel.split_symbol.  Only the adapter and
integration tests have currently been converted, the core/model
tests have not.  You can use the SEQUEL_NO_SPLIT_SYMBOLS
environment variable to disable symbol splitting when running
the adapter and integration tests.