Speed up retrieval of large ado result sets by about 25%
Also, fix handling of fractional seconds when converting timestamp
fields, by converting them to the appropriate timestamp precision.
This only appears to work on ruby 2.2+, probably do to changes in
win32ole between ruby 2.1 and 2.2.  On ruby 1.8, dates and
timestamp values are returned as strings, so deal with that as
well.

This also allows overriding conversion procs on a per-Database
basis.