Make Dataset#freeze actually freeze the dataset on ruby 2.4+
This uses ruby 2.4's clone(freeze: false) support so that you can
clone a dataset that has had several modules extended on it, modify
the opts in the clone, and also extend the object with additional
modules.

This adds the use of initialize_copy for clone to use, which
simplifies clone a bit.  This requires separate implementations of
with_extend, extension, dup, and freeze on ruby 2.4 to deal with
real dataset freezing.

One of the biggest changes here is related to columns, which
was previously stored as an instance variable, and only thread safe
because current ruby implementations make instance variable setting
thread safe (but I don't think that's required).  This changes to
using a plain hash for storing cached data such as the columns.
This updates the prepared statements to also store the prepared
sql in the cache, instead of trying to modify opts.

Now, columns calls columns! if dataset has not cached the columns,
instead of columns! clearing the instance variable and then calling
columns.