Cache SELECT SQL for most frozen datasets
If the dataset is frozen, the SQL produced the last time should be the
same.  There are a couple cases where this may not be true, so handle
those cases by specifically disallowing caching.

Improves Dataset#sql performance by 6x even for the simplest datasets,
with a larger speedup for more complex datasets. On SQLite for a
simple dataset, this translates to a 22% performance increase in
Dataset#all.