Refactor dataset caching slightly
Prefix the symbol cache keys with an underscore, and document
that such keys are only for internal use.

Add cached_dataset private method to DRY up a lot of code where
datasets are cached.

Add _columns and clear_columns_cache method for dealing with
cached columns.

Add disable_sql_caching! private method for disabling SQL
caching for the current dataset, and just use a single
cache_sql? method instead of separate methods per query type.

Don't retrieve cached columns in the current dataset in clone,
first because the result is not used, and second because it
is not thread safe.