Speed up repeated calls to with_pk on the same frozen model dataset using a cached placeholder literalizer
Creating a placeholder literalizer on every call is not worth it,
as in many cases with_pk will be called on a dataset that will be
reused.  This creates a placeholder literalizer on the third call,
so that subsequent calls will be faster.  This provides a 67%
increase in performance in my benchmark using SQLite.