Move identifier mangling code to identifier_mangling extension, load by default unless using :identifier_mangling=>false Database option
Most users aren't going to be modifying the default name mangling, so
it doesn't make sense to support it by default if it can be
implemented via an extension.

This provides a modest speedup for literalizing identifiers, but the
main reason to do it is it makes the code cleaner and simpler.

This removes the need to use ResetIdentifierMangling in all the
shared/sub adapters to make sure the mangling is reset.  Now, the
name mangling extension is not loaded until after adapter
initialization, and during adapter initialization, the default
mangling will always be used.

The default of upcasing input identifiers and downcasing output
identifiers has not changed.

To DRY up adapters that need to not modify adapters (the most
common ones, which all fold unquoted identifiers to lowercase),
add Sequel::UnmodifiedIdentifiers module.