Add Dataset#cached_placeholder_literalizer private method
This method extracts the logic from the Model dataset #with_pk
method, and allows it to be reused for other keys.  It also changes
it to use a single key instead of two, and treats integers as a
counter.  This can result in a cached literalizer being replaced
by a counter in a race condition, but that's not actually a
problem as it would just fall back to the slower code path in
that case.