Speed up repeated calls to Dataset#{first,last} with arguments/blocks for frozen datasets using a cached placeholder literalizer
This removes the previously special handling for last in both
plain datasets and model dataset, which is not needed now that
reverse is cached (though this new code is slightly slower).

For model datasets, this caches the dataset used for ordering by
the primary key, which both last and paged_each can use.

This speeds up calls to first with args or blocks about 25%,
last about 35%, and last for model datasets about 80%.