Cache Dataset#from calls with a single symbol argument in freeze_datasets extension
This provides a decent speedup to Dataset#from by itself, but
the big gain here is that you can use the cache.  For example

  DB[:foo].first(10)

Usually creates a bunch of intermediate datasets and has build
an SQL string everytime.  This makes it create no intermediate
datasets, build a placeholder literalizer (on the 3rd call), and
there doesn't have to build the SQL string everytime, resulting
in about a 35% speedup.