Don't use deprecation warning where :graph_{block,{only_,}conditions} association option is nil
If you don't want to change the default associations behavior,
but just want to avoid the warning, it's better to set one of
these options to nil.  The deprecation warning is there to
protect users who inadvertently call eager_graph/association_join
on a dataset where the :conditions can't be used, but a user
who sets one of the :graph_* options has presumably thought
about that.