Raise Sequel::Error if using an :eager_limit dataset option when eager loading a singular association
The limit for singular associations is always one, so it doesn't
make sense to use the :eager_limit option in that case.  The only
reason to try to support it would be to allow for per-call offsets,
but that would complicate things a little bit and I doubt anyone
needs it.