Handle eager load callbacks correctly for one_to_one associations with orders or offsets when window functions are not supported
The basic issue here is that assign_singular is set statically for
a one_to_one association based on the eager_limit_strategy in play,
but using a dynamic eager load callback could have resulted in a
an eager_limit_strategy being used that was not compatible with
the association's statically defined assign_singular? setting.

Rather than modify the API to support dynamic assign_singular, just
reset the strategy to nil if the strategy is :ruby and
assign_singular is set, so it won't call
apply_ruby_eager_limit_strategy and break things.