Split adapter-dependent code from shared SQLite adapter
The previous commit introduced JDBC specific API code in the shared
adapter.  While it didn't break on other adapters as it used
respond_to?, it's still no the correct place.

Add an sqlite_error_code Database private method that takes the
exception and returns the exception code if there is one, then
define the method appropriately in the jdbc-sqlite subadapter.

I looked and ruby-sqlite3 does not support this currently, though
it offers an API (currently broken) for getting response codes.
Add some commented out code for it in the sqlite adapter, which
can be enabled if ruby-sqlite3 is updated to support the extended
result codes.