Make prepared_statements_associations plugin work correctly when model object explicitly specifies server to use when also using sharding plugin
Without the sharding plugin, association methods are not expected
to use the same shard as the model object, but with the sharding
plugin, they are.  The prepared_statements_associations plugin
wasn't handling that correctly, always using the default shard.

This uses the return super if defined?(super) trick so that the
prepared_statements_associations plugin can integrate with the
sharding plugin without depending on it.