Support Database#freeze
This adds Database#freeze and makes included adapters and
Database extensions implement freeze so that it operates as
expected.

After freezing a database, the database remains usable, but you
can no longer change the configuration, add extensions, etc.

The basic approach is all hashes and arrays used as internal
datastructures are also frozen, unless they are expected to be
modified during use (prepared statements), in which case access
to them must be thread-safe.  Additionally, for any methods that
cache results in Database instance variables, those are called
before freezing to ensure the instance variable already exists
and the method will just return the cached value.

Freezing a Database instance also freezes the related connection
pool.

It will be recommended to freeze all Database instance in production
and when testing, to reduce the risk of thread safety issues and to
easily see if something is attempting to modify the database after
initialization.

External adapters may need to override Database#freeze, freeze
internal structures, do necessary setup, then call super to
correctly support it.