Make Database#extension not attempt to load the same extension more than once
If an extension is already loaded, it doesn't really make sense to
try to load it again.  While this isn't usually a problem to do so,
it generally breaks if the database instance is frozen, since most
extensions load a module into the database.

Technically this iS Not backwards compatible, but hopefully nobody
is relying on the previous behavior