Optimize Model#refresh similar to Model.with_pk
This applies the same optimization used by Model.with_pk to
Model#refresh, taking a standard query and just appending the
primary key value to it. It drops the LIMIT 1, which should
be fine as the primary key should only return a single row.

Most of the commit is spec churn to deal with the simplified SQL.