Make prepared_statements plugin not use prepared statements in cases where it is probably slower
After benchmarking on 6 separate databases, it appears that in
many cases, the prepared_statements plugin actually results in
a net slowdown, due to how optimized the other parts of Sequel
are.

It appears to always make sense to prepare insert and update
statements, as large values may be used in the query.  For
deletes and lookups by primary key, it only makes sense to
do so if Sequel isn't able to use some of its optimizations.
The exception being deletes on Oracle and DB2, which seem
to be faster when prepared statements are used, even considering
Sequel's optimizations.

While here, change things so the prepared_statements plugin does
not add a method to Sequel::Model itself.  Instead, use
defined?(super) to check for an earlier definition, and default
to what the earlier definition returns if it does not return nil.

This adds a class level use_prepared_statements_for_pk_lookup?
method, analogous to the instance level use_prepared_statements_for?
method.  The class method and instance method can be overridden
to return true if you want the previous behavior of always using
prepared statements.