Add Sequel::Postgres.sequel_pg_version_supported? if pg is used
When making a change to Sequel that will make it incompatible
with previous versions of sequel_pg, we can increment the version
here, and then those versions of sequel_pg will no longer break
things.

While here, remove the explicit definition of Sequel::Postgres,
as the shared adapter should define it.