Fix handling of model finders for frozen model classes
When the class is frozen, create finders for all finder_loaders.
Then if the class is frozen, you can access the hash directly
without using a mutex.

Remove the mutex synchronization for finder_loaders in one place,
since the mutex wasn't used in other places.

Correctly use a mutex to synchronize access to finders when
resetting the instance dataset.