Literalize SQLTime instances with only three fractional digits in jdbc/postgresql
This wasn't necessary in JRuby 9.1.7.0, but appears to be necessary
in JRuby 9.1.8.0 to get the tests passing. In JRuby 9.1.7.0, Sequel
used to use 6 fractional digits, but the last three were all zero.
In JRuby 9.1.8.0, the last three are not all zero.

java.sql.time only supports milliseconds and not microseconds, so
trying to use all 6 fractional digits results in a test failure.

This appears to fix the only test failure on JRuby 9.1.8.0, which
was the first JRuby version to support microsecond precision for
Time.