Add Model.dataset_module_class accessor, allowing plugins to support custom behavior in dataset_module blocks
This fixes the issue where eager was a valid method in dataset_module
blocks even if the association support wasn't loaded.  However,
plugins that add new dataset_methods can now also add new helper
methods in dataset_module blocks.

The correct way to do this for an arbitrary plugin is:

  model.dataset_module_class = Class.new(model.dataset_module_class)
  model.dataset_module_class.include(DatasetModuleMethods)

The version used for associations cheats because associations is
always loaded first if it is loaded at all, so it just assigns
a specific subclass.