Handle dumping of autoincrementing 64-bit integer primary key columns correctly when using :same_db option in the schema dumper (Fixes #1324)
When using the :same_db option, Sequel generally uses the actual
database type.  However, at least on PostgreSQL, this causes issues,
because the actual database type (bigint) doesn't match the database
type you specify when creating the column (bigserial).  If you use
the generic :Bignum type, Sequel knows how to translate that
correctly, but specifying the type as a string leads to that string
being used verbatim.

To fix this, if using the :same_db option and the actual database
type matches the 64-bit integer type that Sequel uses for :Bignum
columns, use :Bignum instead of the actual type.