Deprecate dataset mutation
As announced in the 4.42.0 release notes, dataset mutation will
not be supported in Sequel 5.

This deprecation messages are sent when calling raise_if_frozen!,
which now takes an optional method name.  External callers of
raise_if_frozen! should be modified to include the method name
actually called, in order to provide a nicer deprecation warning.