Automatically use from_self when using a dataset as part of a compound if it has an offset but no limit
The situation where the dataset already had a limit but no offset,
or where it had both a limit and an offset, were already handled
correctly.  This allows the only offset case to work.

Also, remove the spec guard for Microsoft SQL Server, since the
related spec now works correctly on it.