Deprecate passing a block to Database#from
For backwards compatibility, a block passed to Database#from is
actually passed to Dataset#where instead of Dataset#from.  This is
different from the behavior of Dataset#from, which uses the block
as a virtual row block.  This behavior is counter intuitive and
makes it more difficult to use set returning functions.  The
from_block extension was added in 4.2.0 to make Database#from
operate like Dataset#from, this will become the default behavior
in Sequel 5.