Deprecate support for some databases and adapters
This deprecates all support for Cubrid, Firebird, Informix, and
Progress databases.  Informix and Progress had barely any support.
Firebird has some support, but it couldn't even run the tests.
Cubrid was the best of the four, but had many test failures.

This also deprecates the do (DataObjects), swift, and jdbc/as400
adapters.  While the do and swift adapters work OK, there are few
users, and it doesn't seem worthwhile to maintain support.

In all of these cases, users who are still using any of these
adapters can continue to do so, they just need copy the existing
code and move it to an external adapter that they maintain
themselves.