Add deprecation message for using association_pks setter method without :delay_pks association option
Also add one for when :delay_pks=>true is used.

The default in Sequel 5 will be :delay_pks=>:always.  If you want
the current behavior, you'll have to use :delay_pks=>false. There
will be no setting for the current :delay_pks=>true behavior,
because it is inconsistent to only delay for new objects and not
for existing objects.  However, we can consider a
:delay_pks=>:only_new or similar option if someone really needs the
current :delay_pks=>true behavior.