Deprecate calling Dataset#{update/delete/truncate} on datasets with limits or offsets unless the database supports it
Sequel's historical default behavior when doing:

  DB[:table].order(:foo).limit(2).delete

Has been to ignore the order and limit and delete the entire
table.  On MySQL the behavior has been to use an order and limit
for deletes and updates, and on Microsoft SQL Server the behavior
has been to use a limit for update but not for delete.  I'm not
sure why delete with a limit wasn't supported before, since
Microsoft SQL Server has supported it since 2005, so I added
support for it.

The current behavior can break POLS for users who do not understand
that databases generally don't support limited updates/deletes.  Also,
there are numerous corner cases even on databases that do support
them. MySQL supports limits but not offsets, and Microsoft SQL Server
supports limits but only if unordered, and also does not support
offsets.

With this change, Sequel will now issue a deprecation warning if
a limit or offset is specified but ignored.  Also, on Microsoft
SQL Server, Sequel will now issue a deprecation warning if
specifying a limit and an order, since the order would be ignored.

One of the reasons this wasn't done previously is that certain
parts of Sequel rely on updating limited datasets.  In some cases
the limits could be removed, but behavior changes are less likely
if there is just a flag to keep the existing behavior for those
datasets.  Dataset#skip_limit_check has been added for that
purpose, and is used automatically by Sequel::Model in certain
cases so that normal model usage will work in Sequel 5.