Add validation_contexts plugin for supporting custom contexts when validating
This is modeled somewhat on the similar feature in ActiveRecord.
Sequel::Model#save/valid? already take an options hash, so this
is just added as another option.  Sequel::Model#validate is
designed to be overridden by the user, so that API can't be changed.
So this just sets the validation context to an instance variable
before calling validate, and sets it back to nil afterward.  This
approach doesn't work with frozen objects, but is otherwise nicer
if you have custom validation methods where you don't want to
thread the context through multiple methods.