Deprecate passing Procs as filter arguments, require they be passed as blocks
The problem with allowing Procs as filter arguments is that they
delay evaluation, but require additional introspection to
determine if evaluation is delayed.  This complicates some of
code Sequel uses to determine what operations can be optimized,
and makes it not possible to optimize certain operations.

Another problem with allowing Procs as filter arguments is that
they can more easily lead to unbound or unexpected recursion.