Deprecate returning false from model before hooks to cancel the action, use Model#cancel_action
Historically, Sequel supported returning false in before hooks
for canceling the action. The main issue with this is that it
is nonobvious and also possible to accidentally trigger if you
aren't in what the before hook methods return.

Model#cancel_action was added in 4.19.0 as an alternative to
returning false, and will be the supported method in Sequel 5.

This also requires changes to hook_class_methods and
instance_hooks, both of which internally had similar short
circuiting when a before hook returned false.

While making these changes, I found that Model#cancel_action
did not work correctly when using Model#valid?.  This includes
changes to that the two work together.