Deprecate Sequel::PG_NAMED_TYPES
This was previous used to handle named types on PostgreSQL, where
the type oids could differ per database.

Database#add_named_conversion_proc was added in 4.21.0, which
handles this in a significantly better way, so just recommend
that instead.

The pg_hstore extension used to modify PG_NAMED_TYPES, but this
removes the support for that.  Any database that uses the
pg_hstore extension will have support for the hstore type added
when the extension is loaded.