Deprecate the schema and scissors plugins
The schema plugin is bad as it basically is designed around
cases where the model class is created before the table.

The scissors plugin is dangerous.  It was added in Sequel 4 for
backwards compatibility with Sequel 3, but it probably shouldn't
be used as it is too easy to make a mistake using it.