Correctly handle datasets with offsets but no limits used in compound datasets on MSSQL <2012
This was handled correctly on MSSQL 2012+ as offsets don't need to
be emulated in that case.  It seems pretty stupid for the database
to raise an exception for a dataset without a limit where it can
perform the same exact operation if TOP (100) PERCENT is specified,
but this workaround is currently required on old MSSQL versions.