Deprecate splitting symbols to look for qualified/aliased identifiers
Splitting symbols has been around since the first code commit.
While it allows more terse code, it is non-intuitive, causes a lot
of extra complexity, and doesn't work well when dealing with
databases that use double/triple underscores in identifiers, or
leading/trailing underscores in identifiers.

With this commit, symbol splitting is on by default, but a
deprecation message will be issued when splitting a symbol
where the behavior would be different than if it wasn't split.

In Sequel 5, symbol splitting will be off by default, but there
will be a switch to turn it back on.