Actually deprecate Sequel::Postgres::PG_NAMED_TYPES constant
This deprecation was already announced, but only partially
implemented.  Previously, users would get a deprecation warning
if they added a named type and then tried to get a Database
instance to use it.  This makes it so any reference to
Sequel::Postgres::PG_NAMED_TYPES will print a warning, at
least on Ruby 2.3+.