Deprecate Model.def_dataset_method and Model.subset, move to def_dataset_method plugin
These methods both predate the use of dataset_module, but it makes
much more sense for them to be dataset_module methods, since they
are adding methods to the dataset and not to the model itself.

This changes Model.subset and DatasetModule#subset so that
Model.subset calls DatasetModule#subset instead of the reverse.
The previous behavior was a workaround because plugins couldn't
affect dataset_module behavior.  With the recent addition of
dataset_module_class, plugins can now affect dataset_module
behavior, and the inverted_subsets and subset_conditions plugins
now do so.