Deprecate using a joined dataset as a Sequel::Model dataset
Joined datasets cause all types of problems.  Some of the most
significant have been handled, but it's probably best to avoid
the complexity completely.

This makes it so you need to wrap the joined dataset in a subquery,
then you can use it as a model dataset.  Sequel 5 will automatically
use a subquery, using the first table name as the subquery alias.