Don't issue deprecation warnings on ruby 1.8.7, as Sequel 5 is dropping support for it
If you are still using ruby 1.8.7, there is no point in issuing
deprecation warnings, since you can't upgrade to Sequel 5 in that
case.  Anyone using ruby 1.8.7 will be stuck using the last
Sequel 4 release, and then it isn't a problem if they are using
code that will be removed in Sequel 5.