Don't quote function names that are SQL::Identifiers, unless SQL::Function#quoted is used
With the deprecation of virtual row method block support, developers
are encouraged to switch from:

  where{func{} > 1}

to:

  where{func.function > 1}

Unfortunately, on databases that support quoted function names,
this change starts quoting the function names, which can break if
the incorrect case is used.  Change things so that the function
remains unquoted in this case, unless SQL::Function#quoted is
explicitly called.

This keeps the previous behavior for SQL::QualifiedIdentifier
function names, which makes things a bit inconsistent.  However,
users using schema qualified function names are more likely to be
sensitive to quoting, so it's probably better to keep the previous
behavior in that case.