Respect per-database timezones when doing datetimeoffset conversion
Sequel.datebase_to_application_timezone is an old API that doesn't
respect per Database timezones, Database#to_application_timezone
is the appropriate conversion method.

Also, handle case where an old SQL Server JDBC driver is used and
Java::MicrosoftSql::Types::DATETIMEOFFSET is not defined.

Add appropriate spec guards, as datetimeoffset conversion is not
handled correctly on odbc, ado, and older versions of tiny_tds.

While here, make spec for time also do a cast to ensure that it
is correctly interpreted by the database.  This also causes it
to fail on ado for the same reason that datetimeoffset fails on
ado.