Deprecate automatically treating plain strings passed to filtering/update methods as literal SQL
Automatically treating plains strings passed to filtering methods
as literal SQL is the most common cause of SQL injection in
applications using Sequel.  By requiring literal SQL use to be
confirmed explicitly using Sequel.lit, it reduces the likelihood
of SQL injection, and in the case where Sequel.lit is used,
security auditing is much easier.

This behavior has been available in the form of the
no_auto_literal_strings extension for over a year, and I think
it is a better default.

The deprecation message for using placeholder literal strings
says to convert:

  ds.where('a = ?', 2)

to:

  ds.where(Sequel.lit('a = ?', 2))

Unforuntately, that previously did not use parentheses around
the expression.  Start automatically using parantheses around
the expression so that the two result in the same behavior.

This could potentially be not backwards compatible, but you
would have to be relying on implicit operator precedence and
I would consider that exploiting a bug.  For example:

  ds.where(Sequel.lit('? OR ?', :a, :b)).where(:c)
  # Before: WHERE a OR b AND c
  # Now: WHERE (a OR b) AND c

For people who don't want to update their code, the
auto_literal_strings extension has been added to provide the
old behavior.

Using plain strings as literal SQL in the database schema methods
(e.g. for constraints) is still allowed, in order to not break
old migrations.

One other small change related to this is that the following
code is now deprecated, even though it contains a literal string:

  ds.where([Sequel.lit('a')])
  ds.where([Sequel.lit('a = ?'), 'b'])

This type of code should be changed to:

  ds.where(Sequel.lit('a'))
  ds.where(Sequel.lit('a = ?', 'b'))