Deprecate some corner case argument handling for filtering methods
The historical handling of these corner cases is bizarre. Consider:

  dataset.where(nil, &foo)

If foo is nil, and there is no existing where clause, this does
nothing.  If foo is not nil, this adds an AND NULL filter. If
there is an existing where clause, it also adds AND NULL filter.

You could call where with no arguments, which was treated the
same as calling it with an empty array argument.  It doesn't make
sense to support that now that the submitting multiple separate
arguments in an array is already deprecated.

You could call where with an empty string, which would be ignored,
except if you also passed a block.  This was also true for other
arguments that responded to empty?.

Try to deprecate all of this corner case behavior, and in Sequel 5:

* Either argument or block is required
* Empty array or empty hash is ignored, even if a block is passed
* Explicit nil is always treated as NULL