Replace Dataset#_filter_or_exclude with add_filter
_filter_or_exclude was a private method that handled the internals
filtering.  This replaces it with add_filter, which doesn't take
a split, reorders the arguments, and takes an argument for how
to combine with an existing argument.

This changes Dataset#or so that instead of having its own code, it
can now use add_filter.

This should slightly improve performance by avoiding an array
allocation per filtering method call.