Deconstantize dataset SQL generation, speeding up ruby 2.3+, slowing down earlier versions
This basically reverts 3f60e836e86fd8a3129f70e8ddb181c5fd7dbd0d and
29f2a48ac75e7aa70332b22fbe4b415e20ebbbbc, going back to using plain
strings instead of constants when generating SQL.  The reason for
this is that Sequel uses frozen string literals on ruby 2.3+, and
using a frozen string literal is faster than using a constant
reference.

Unfortunately, ruby <2.3 cannot benefit from this, and as such
this change will slow things down on those ruby versions.  I
consider that acceptable.  If you want the fastest performance, run
the most recent version of ruby.  Both MRI 2.3+ and JRuby 9.1+
support frozen string literals.

It would be possible to use frozen string literals on ruby 2.1+
by adding .freeze after every string, but I decided against that
as it really clutters the code.

This deprecates the constants used in SQL generation that are no
longer used, as well as some other constants found when making
these changes.

This also has some other small performance enhancements, such
as combining multiple strings into a single string, and the
use of strings instead of regexps for gsub calls that don't
need a regular expression.  In addition, it adds a lot of
comment reminders to freeze things in Sequel 5.