Added test for incorrect error logging in `table_exists?`
This PR adds the missing spec mentioned in #1356. It doesn't attempt to _fix_ the problem, but it illustrates how it can easily be reproduced.

Note: I haven't really used minitest before; it might be that this can be more elegantly described in some other way. (Much more familiar with `rspec`, which has a very different world view with regards to mocking.)

(The whitespace diffs came as my editor is set to trim trailing whitespace on file save. Can strip away that part of the change if you like; I included it for now since it makes sense.)