Deprecate setting invalid datasets for models unless required_valid_table = false
During the testing for this deprecation, I noticed that require_valid_table
would raise for things like using subqueries.  That wasn't intentional, it
was because the code raised an exception if getting the schema failed, instead
of falling back to the columns in that case.  Now the error is only raised
if required_valid_table is true and the fallback to check the columns also
raises an exception.