Fix incorrect SQL used for inserting into a CTI subclass sharing the primary table when using the :alias option
In a hierarchy such as:

class Employee
  plugin :class_table_inheritance
end

class Manager < Employee
end

Where there is not a separate table for manager data, the INSERT
SQL was incorrect if the :alias option was used.  Previously, it
tried to insert into a subquery, which doesn't make sense.

This checks for whether the primary class was used, instead of
only the primary table.